python/cli/pyproject.toml
[tool.poetry]
name = "data-platform-workflows-cli"
# Version unused; repository has its own versioning system. (See .github/workflows/__release.yaml)
version = "0.1.0"
description = ""
license = "Apache-2.0"
authors = ["Carl Csaposs <carl.csaposs@canonical.com>"]
readme = "README.md"

[tool.poetry.scripts]
redact-secrets = "data_platform_workflows_cli.redact_secrets:main"
collect-snap-bases = "data_platform_workflows_cli.craft_tools.collect_platforms:snap"
collect-rock-bases = "data_platform_workflows_cli.craft_tools.collect_platforms:rock"
collect-charm-platforms = "data_platform_workflows_cli.craft_tools.collect_platforms:charm"
release-snap = "data_platform_workflows_cli.craft_tools.release:snap"
release-rock = "data_platform_workflows_cli.craft_tools.release:rock"
release-charm = "data_platform_workflows_cli.craft_tools.release:charm"
update-bundle = "data_platform_workflows_cli.update_bundle:main"
parse-snap-version = "data_platform_workflows_cli.parse_snap_version:main"
allure-add-default-for-missing-results = "data_platform_workflows_cli.allure_add_default_for_missing_results:main"
add-ssh-keys = "data_platform_workflows_cli.add_ssh_keys:main"
tee-log-for-all-models = "data_platform_workflows_cli.tee_log_for_all_models:main"
sync-docs = "data_platform_workflows_cli.sync_docs:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
